requirements.txt
streamlit
google-genai
joblib
scikit-learn
python-whois
requests
python-dotenv
flask
twilio
